migrate query integration tests to `apstra_test` package
